Promote `api` command to a `pkg/cmd/api` package
